delete usuario

+++ src/main/java/hus/censoCamas/security/repo/UsuarioRepo.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.Optional;
 
 public interface UsuarioRepo extends JpaRepository<Usuario, Integer> {
-    void deleteUsuarioById(Long id);
+    void deleteUsuarioById(Integer id);
 
     Optional<Usuario> findByUsuario(String usuario);
